internal/store: remove identity UUID
The identity manager previously generated a UUID for each identity
based on the username. As we would like to make it so that it is
possible for usernames to change this is not a good choice of UUID
scheme. Unfortunately because of the way MongoDB 2.4 upserts work this
UUID had to be deterministic as it is not possible to set the _id field
using $setOnInsert. The UUID however is not actually used in any other
part of the system and MongoDB is more than capable of creating a unique
_id for its own purposes.