idp: rework IDP interface
Rework the IDP interface to enable some changes that will be needed for
the azure IDP. This adds support for Init function in the IDP so that it
can do some setup within its context. The idp.Context type is updated
to also be a context.Context for convenience. The IdentityProvider.URL
method can no longer return an error, which means that the UbuntuSSO IDP
has changed implementation slightly. The WaitID now has to be passed in
to Context methods that need it as it cannot be present as a form field
in the OpenID Connect login flows.