internal/v1: change agent creation API
This removes the old PUT endpoint for updating user information
and replaces it with a POST endpoint that creates a new agent
and chooses its agent name too.

This is a backwardly incompatible change, but this endpoint
was previously only available to admin clients and the
only users we know of are:

- the user-admin command
- the blues-identity charm

The former is updated by this PR; I don't think we need to worry
about existing scripted uses of it.

The latter means that the charm will need to be updated at the same time as the new idm version is deployed. If that's a problem, we could reinstate the old endpoint, at least partially.