idp: do not use Groups field to store idp-specific groups
The Identity.Groups field is currently being used by some identity
providers as a cache of group information retrieved at login time.

Doing this is problematic because the Groups field is also used
for groups added locally via the API and there's no easy way
for a provider to tell the difference, so the groups are being added
to but not removed.

This PR changes the providers so that they store group information
in ProviderInfo instead.

This will require a database migration to move information out of Groups and into
ProviderInfo.