internal/discharger: split legacy discharge endpoints
This makes the division between legacy and non-legacy protocols
clearer. We also make the tests test a bit more of the logic
via the bakery.

Also use httpbakery.ThirdPartyCheckerP so we have access
to the response for setting cookies.

We don't actually set the cookie yet because there are
some ongoing security thoughts about that need to be
resolved first.