fix candid command agent interaction
We were using web-browser interaction in preference to
agent interaction, so it was failing.

Change the logic so that if an agent file is specified, we'll
avoid web-browser authentication entirely and
also avoid using cookies.

Also clarify the candid command help a little and make
it an error to specify groups in put-agent when they're
going to be ignored.