storage plugin architecture
We use a similar approach for storage in the config as we
are currently doing for identity providers. Each storage
backend is responsible for registering itself.

So that everything doesn't depend on the config package,
idp implementations now register themselves with the idp
package and storage backends register themselves with the
store package.

We also move the store backend packages under the store
package so that they're in the same structure as the idp implementations.