internal/discharger: use random discharge ids
Previously, if two clients attempted to discharge
a macaroon with the same third party caveat id at the
same time, one of the attempts would fail because
discharge ids were deterministically determined from
the caveat id.

Since we no longer tie discharge tokens to a particular
discharge, we no longer need to know the discharge id
in advance, so we remove the for-discharge caveat and
make the discharge id random for every new discharge
request.