Rework candid snap
This change removes the separate candid and candidsrv snaps and matching
Makefile targets, instead replacing them by a single snap called
"candid" which provides both the client and server.

Go 1.11 is used for building Candid and the build has been tested on
Launchpad builders using a core16/16.04 base.

This snap also generates a default config on initial startup, including
a new key pair and defaults to a working setup using non-persistent
storage and a static identity provider.

This change also moves the snapcraft.yaml to the top-level directory so
that the Git repository may be used with build.snapcraft.io.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>