internal/discharger: change short-term stores
Change the types stored in the short-term data stores used in the login
protocol to store an Identites unmodifiable ProviderID rather that a
DischargeToken. DischargeTokens are now only created when they are
needed. This is a preliminary branch to support using ProviderIDs in the
discharge tokens.

Please note that, while this change is isolated to the discharge
package, some in-progress logins that are happening at the time of
upgrade to this code may fail. If logins do fail then they will work
if retryed immediately afterwards.