cmd/candidsrv: fix http-proxy configuration
Go 1.16 changed how the HTTP_PROXY environment variable is interpreted
such that it is no longer used for HTTPS connections. The handling for
this setting has therefore been changed to modify the DefaultTransport
rather than the environment.

Also update the core snap used in candid to core18 as 16.04 is going
end-of-life.